SRC/requirements.txt
pandas==1.3.2
numpy==1.21.2
scipy==1.7.1
seaborn==0.11.2
matplotlib==3.4.3
plotly==5.3.1
wordcloud==1.8.1
sklearn==0.0
IPython.display==7.27.0
bs4==0.0.1
requests==2.26.0
pytest==5.4.3
pickleshare==0.7.5
pandas-profiling==3.1.0
pycaret==2.3.4
pycaret-nightly==2.3.4.dev1637109959
sklearn==0.0
catboost==1.0.3
glob2==0.7
scikit-image==0.18.3
scikit-learn==0.23.2
scikit-plot==0.3.7
opencv-python==4.5.4.58
tf-estimator-nightly==2.8.0.dev2021112009
tf-nightly-gpu==2.8.0.dev20211120
